+++ MavenDemo/src/main/java/cn/taike/design/mode/DesignMode.java
@@ -1,0 +1,11 @@
+package cn.taike.design.mode;
+
+/**
+ * Created by huayandong on 17/8/9.
+ */
+public class DesignMode {
+
+    public static void main(String[] args) {
+        System.out.println("aaa...");
+    }
+}
