+++ MavenDemo/src/main/java/cn/taike/design/mode/command/CommandOn.java
@@ -1,18 +1,18 @@
-package cn.taike.command;
+package cn.taike.design.mode.command;
 
 /**
  * Created by huayandong on 17/8/2.
  */
-public class CommandOff implements Command {
+public class CommandOn implements Command {
 
     private Light light;
 
-    public CommandOff(Light l) {
+    public CommandOn(Light l) {
         this.light = l;
     }
 
     @Override
     public void execute() {
-        light.turnOff();
+        light.turnOn();
     }
 }
