+++ MavenDemo/src/main/java/cn/taike/entity/Book.java
@@ -2,27 +2,18 @@ package cn.taike.entity;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 
-import javax.persistence.*;
-
 /**
  * Created by huayandong on 17/7/24.
  */
-@Entity
-@Table(name = "book")
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 public class Book {
 
-    @Id
-//    @GeneratedValue(strategy = GenerationType.AUTO)
     private Integer id;
 
-    @Column(name = "book_name")
     private String bookName;
 
-    @Column(name = "project_name")
     private String projectName;
 
-    @Column
     private String type;
 
     private String cover;
