+++ MavenDemo/src/main/java/cn/taike/jdbc/template/JdbcSample.java
@@ -1,4 +1,4 @@
-package cn.taike.dao;
+package cn.taike.jdbc.template;
 
 import java.sql.*;
