--- MavenDemo/src/main/java/cn/taike/schedule/Job.java
@@ -1,12 +1,0 @@
-package cn.taike.schedule;
-
-/**
- * Created by huayandong on 17/7/14.
- */
-public class Job {
-
-    //任务类
-    public void doJob() {
-        System.out.println("定时任务在执行...");
-    }
-}
